requirements.txt
# requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
pillow==10.1.0
celery==5.3.4
redis==5.0.1
channels==4.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
django-extensions==3.2.3
django-filter==23.5
django-crispy-forms==2.1
crispy-bootstrap5==2.0.2